requirements.txt
python-telegram-bot==20.7
fastapi==0.108.0
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.6.1
pydantic==2.5.2
pydantic-settings==2.1.0 